Ignoreing PHPStorm and moving all Api Errors to new namespace.
