+++ .gitignore
@@ -1,5 +1,6 @@
 .DS_Store
 
+.idea
 composer.phar
 composer.lock
 /node_modules/
