+++ lib/Castle/Errors/Api/ApiError.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
+
+use Castle\Errors\CastleException;
 
 class ApiError extends CastleException
 {
