+++ lib/Castle/Errors/Api/ChallengeRequiredError.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
 
 class ChallengeRequiredError extends ApiError
 {
