+++ lib/Castle/Errors/Api/ForbiddenError.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
 
 class ForbiddenError extends ApiError
 {
