+++ lib/Castle/Errors/Api/InvalidParametersError.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
 
 class InvalidParametersError extends ApiError
 {
