+++ lib/Castle/Errors/Api/NotFoundError.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
 
 class NotFoundError extends ApiError
 {
