+++ lib/Castle/Errors/Api/UnauthorizedError.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Castle\Errors;
+namespace Castle\Errors\Api;
 
 class UnauthorizedError extends ApiError
 {
